refactor: notification batch module
