+++ sseudam-core/core-application/src/main/kotlin/com/sseudam/notification/NotificationFacade.kt
@@ -4,6 +4,7 @@ import com.sseudam.notification.command.CreateNotificationStoredCommand
 import com.sseudam.notification.command.FirebaseCloudMessageCommand
 import com.sseudam.notification.component.NotificationStoredKeyGenerator
 import com.sseudam.notification.dto.NotificationMessages
+import com.sseudam.notification.fcm.FcmSender
 import com.sseudam.user.UserDeviceService
 import com.sseudam.user.UserService
 import org.springframework.stereotype.Service
@@ -14,7 +15,12 @@ class NotificationFacade(
     private val userService: UserService,
     private val notificationStoredKeyGenerator: NotificationStoredKeyGenerator,
     private val notificationService: NotificationService,
+    private val fcmSender: FcmSender,
 ) {
+    companion object {
+        private const val DEFAULT_USER_NICKNAME = "사용자"
+    }
+
     fun createWeeklyNotificationMessages(): List<FirebaseCloudMessageCommand> {
         val userDevices = userDeviceService.findAll().filter { it.fcmToken.isNotBlank() }
         if (userDevices.isEmpty()) {
@@ -55,5 +61,48 @@ class NotificationFacade(
         )
 
         return messages
+    }
+
+    fun sendNewPetNotifications() {
+        val userDevices =
+            userDeviceService
+                .findAll()
+                .filter { it.fcmToken.isNotBlank() }
+        if (userDevices.isEmpty()) return
+        val userProfiles =
+            userService
+                .findAllBy(userDevices.map { it.userId }.distinct())
+                .associateBy { it.id }
+        val messages =
+            userDevices.map { device ->
+                FirebaseCloudMessageCommand(
+                    fcmToken = device.fcmToken,
+                    title = NotificationMessages.DEFAULT_TITLE,
+                    body =
+                        NotificationMessages.newPetContents(
+                            userProfiles[device.userId]?.nickname
+                                ?: DEFAULT_USER_NICKNAME,
+                        ),
+                )
+            }
+        fcmSender.sendAll(messages.toSet()).apply {
+            notificationService.appendAll(
+                messages
+                    .map { message ->
+                        CreateNotificationStoredCommand(
+                            userId =
+                                userDevices
+                                    .find { it.fcmToken == message.fcmToken }
+                                    ?.userId
+                                    ?: return@map null,
+                            notificationStoredKey = notificationStoredKeyGenerator.generate(),
+                            type = "PET_SEASON",
+                            parameterValue = "",
+                            topic = message.title,
+                            contents = message.body,
+                        )
+                    }.filterNotNull(),
+            )
+        }
     }
 }
